마케터 세금계산서] 주석제거, makeQuery함수 util로 이전
